refactor: extract selection state management into useRatesSelect hook
